[Go-Builder] Packages can contain any file and refer to them any time
This commit renames the file from the nexus package, without changing
the contents and how they refer to files. In addition adding the method
that verifies if the files exists before downloading it. Since the
release versions never change, no need to verify caching info.

The biggest change is that all the nexus-related metadata was moved to
the metadata module. Everything related to the service itself was moved
to the service module.

*	new file:   nexus/metadata.go
- Separating the nexus metadata structs and parsing of the xml from the
  downloaded file metadata.

*	renamed:    nexus/nexus.go -> nexus/service.go
- Renaming the service-related classes, reusing the metadata classes.
- No need to import as they are in the same "nexus" package.