[Go-Builder] Scrapping the list of resources from a package, using Closures
This commit makes a major refactoring in the nexus package to provide
the list of resources from a given artifactId. The terms have changed
and the nexus package implementing a way to filter Nexus artifacts.

The main go file was temporarily commented out for the verification
of the output. The list is filtered with the use of a Closure type.

*	modified:   main.go
- Commenting out the cfp service for now until integration with
  Nexus package.

*	modified:   nexus/metadata.go
- The metadata refactoring removing the constant

*	new file:   nexus/artifacts.go
- The representation of artifacts for nexus.
- For now, just the artifacts list.