[Nexus Downloader] Refactoring the progress output: Resource name adjusted + Last Download to 100%
This commit refacts the download and progress to properly handle the
last download by waiting for 1sec after the process finishes. It
properly fills out the total download.

Finally, changing the progress bar library to use the forked version
because the current version imposes the locked-terminal. Until it is
resolved in the master branch, I will keep my fork.

https://github.com/cheggaaa/pb/issues/31

*	modified:   nexus/download.go
- Removing the dependency to the progress libraries, refactored to the
  progress.go.
- Retrieving the instance of the pool
- Waiting 1sec to finish the download

*	modified:   nexus/progress.go
- Using my personal forked version of the PB library until the
  locked-terminal version gives the API to choose to not block the
  terminal.
- Updating the docs
- Fixing the prefix to be "[ obj.ext ]"
- Creating factory methods to properly create the library's version.
- Creating the pool based on the parameters.